requirements.txt
mcp
httpx
pydantic
python-dotenv
uvicorn
fastmcp
logfire
rich
starlette
